refactor => one dir for all components, one dir for layouts using Components
