+++ src/Components/Gallery/Card/Card.css
