+++ src/Components/Gallery/Card/Card.js
