--- src/Components/Main/Main.js
@@ -1,14 +1,0 @@
-import React from "react";
-import "./Main.css" ;
-
-class Main extends React.Component{
-    render(){
-        return (
-            <main>
-
-            </main>
-        )
-    }
-}
-
-export default Main ;
