+++ src/Layouts/Header/Header.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "./Header.css" ;
-import Logo from "./Logo/Logo" ;
-import Navigation from "./Navigation/Navigation";
+import Logo from "../../Components/Logo/Logo" ;
+import Navigation from "../../Components/Navigation/Navigation";
 
 class Header extends React.Component{
     render() {
