+++ src/Layouts/Main/MainHome.js
@@ -1,0 +1,18 @@
+import React from "react";
+import "./MainHome.css" ;
+
+class MainHome extends React.Component{
+    constructor(props){
+        super(props)
+    }
+
+    render(){
+        return (
+            <main>
+
+            </main>
+        )
+    }
+}
+
+export default MainHome ;
